clean more

+++ tests/pipelines/spectrogram_diffusion/test_spectrogram_diffusion.py
@@ -18,7 +18,6 @@ import unittest
 
 import numpy as np
 import torch
-import os
 
 from diffusers import SpectrogramDiffusionPipeline
 from diffusers.utils import slow, require_torch_gpu, torch_device
@@ -61,7 +60,7 @@ class PipelineIntegrationTests(unittest.TestCase):
     def test_spectrogram(self):
         device = torch_device
 
-        pipe = SpectrogramDiffusionPipeline.from_pretrained("/home/patrick/t5_note")
+        pipe = SpectrogramDiffusionPipeline.from_pretrained("/home/patrick/t5_note", torch_dtype=torch.float16)
         pipe = pipe.to(device)
         pipe.set_progress_bar_config(disable=None)
